+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 - Change width of items in status bar
 - Reduce pattern drawing cost in playback
+- Improve SSG tone/noise editor
 
 ## v0.3.0 (2019-10-25)
 ### Added